refactor(chat): collapse getChatMcpTools tool wrappers into chat-tool-builder (#5528)
Stage B of the AI-SDK unification (stacked on #5526).

## What

- Characterization tests (`chat-mcp-client.tools.test.ts`) pinning
`getChatMcpTools` behavior before the restructure: per-kind tool shape
(description fallbacks, `toModelOutput` on MCP tools only, schema
normalization), the MCP execute pipeline (pre-hook → gateway call →
post-hook order, pre-hook block short-circuit, post-hook feedback
append), the agent-delegation execute path (previously zero coverage),
approval gating (`blockOnApprovalRequired`, `run_tool` grant approval,
no retargeting for delegation tools), error paths, metric emission
counts, and tool-cache gating (`abortSignal` bypass, headless reuse,
per-conversation scoping).
- The two near-identical inline AI-SDK tool wrapper blocks inside
`getChatMcpTools` (MCP gateway tools and agent delegation tools, ~430
lines combined) collapsed into `buildMcpGatewayTool` /
`buildAgentDelegationTool` plus shared helpers (`needsApprovalProps`,
`executeWithToolSpan`) in the new `chat-tool-builder.ts`. Execution
internals (`executeMcpTool`, hook helpers, approval helpers, schema
normalization, metrics) moved there verbatim. `chat-mcp-client.ts` keeps
token selection, client lifecycle, caches, and the unchanged public API
(2334 → 1035 lines); import direction is one-way into the builder
module.
- Dropped the unused `__test.pingClientWithTimeout` export.

No externally observable behavior change: tool
names/descriptions/schemas, outputs, error shapes, approval flows, hook
semantics, metrics, logs, and cache keying/gating are preserved. Cache
semantics (including the `!abortSignal` gating and what the cache
stores) are intentionally untouched in this stage.

## Notes

- Delegation `needsApproval` now routes through
`resolveApprovalPolicyTarget`; equivalent for every name `getAgentTools`
can produce (`agent__*` never parses as `run_tool`), and a test pins
that delegation approval does not retarget via `tool_name` in args.
- Test-file mocks sit at process boundaries only: MCP SDK transport,
`mcpClient.executeToolCall`, `executeA2AMessage` (child-agent run),
`hookDispatcherService.fire` (hook scripts run in Dagger sandbox
containers), browser-stream feature, external-IdP token resolver.

## Validation

- backend: 7967 passed / 8 skipped (full suite), type-check, biome, knip
dev+production all clean
- workspace `pnpm type-check` and `pnpm lint` clean

---
<!-- archestra-banner:v1 -->
<a href="https://archestra.ai/contributor-onboard" rel="nofollow
noreferrer noopener" target="_blank">

<img alt="Archestra Contributor"
src="https://raw.githubusercontent.com/archestra-ai/archestra/main/docs/assets/archestra-contributor-banner.webp"/>

</a>
